refactor(throughput): give the probe's parts types and files of their own
base.rs had grown to 639 lines of mostly free functions over two global
statics, with nothing saying which of them belonged together or which
were the entry points.

Five modules, each a type that owns its own state and its own tests:

- `pooling`: `PooledProbes` and the map of who is holding a device's pool.
- `launch`: `LaunchConfig`, built by `for_device` and resized by `with_units`.
- `shape`: `ShapeSweep`, ranking the shapes a probe can be launched in.
- `workers`: `WorkerSweep`, the CPU worker sweep and its per-device memo.
- `operands`: `Arithmetic` and `CooperativeMatrix`, what a compute probe
  can be measured over.

base.rs keeps the public API and the dispatcher that reads the key and
picks a runner, which is what a reader opens it for.

Both statics were file-scoped globals reachable from anywhere in those
639 lines and are now private to the module holding the only type that
touches them.
