+++ crates/cubecl-std/src/throughput/base.rs
@@ -1,40 +1,19 @@
-use cubecl_common::device::ServiceId;
-use cubecl_core::ir::{ElemType, FloatKind};
-use cubecl_environment::{collections::HashMap, sync::Mutex};
+use cubecl_core::ir::ElemType;
 use cubecl_runtime::{
     client::Client,
     runtime::Runtime,
-    server::CubeDim,
     throughput::{
-        ComputeCmmaConfig, KernelConfig, MemoryAccess, MemoryCurve, MemoryPoint, MemorySpec,
-        ThroughputBenchmarker, ThroughputError, ThroughputKey, ThroughputMode, ThroughputValue,
-        sweep_size, working_set_sweep,
+        MemoryAccess, MemoryCurve, MemoryPoint, MemorySpec, ThroughputError, ThroughputKey,
+        ThroughputMode, ThroughputValue, sweep_size, working_set_sweep,
     },
     tune::{Bounds, Thresholds, Work, calculate_bounds},
 };
 
 use crate::throughput::{
+    Arithmetic, CooperativeMatrix, LaunchConfig, PooledProbes, ShapeSweep, WorkerSweep,
     compute_cmma, compute_direct, launch_overhead, memory_direct, memory_probe, memory_read,
     memory_write,
 };
-
-/// Independent cube positions each CPU worker interleaves, so a compute
-/// pass pipelines past instruction latency instead of serializing on one
-/// dependency chain. A depth, not a machine guess: a handful hides any
-/// core's fma latency, excess is free because the iteration budget is
-/// time-calibrated, and nothing about the launch scales with it. Memory
-/// probes with blocked addressing pin their own count back to one (see
-/// [`MemoryProbe::new`](crate::throughput::memory_probe::MemoryProbe::new)).
-const CPU_CHAIN_DEPTH: usize = 64;
-
-/// Worker counts a memory probe is swept over before the fastest is kept, once
-/// per device and access. Five halvings reach a sixteenth of the cores, and
-/// each one costs a ranking pass.
-const MEMORY_WORKER_SHAPES: usize = 5;
-
-/// Units a GPU probe asks for. A wider cube measures no faster, and makes the
-/// memory probes report several times the bus rate.
-const PROBE_UNITS_PER_CUBE: u32 = 256;
 
 /// Measure peak throughput on `device` for each of the given `keys`.
 pub fn device_throughput<R: Runtime>(
@@ -51,9 +30,8 @@ pub fn device_throughput<R: Runtime>(
 /// kilobytes up to as much as the device will allocate.
 ///
 /// One point per size in [`working_set_sweep`], each measured and cached
-/// exactly like the single-size probe — [`measure_peak_throughput`] with a
-/// [`ThroughputMode::Memory`] key — so a curve costs one probe per size on the
-/// first run and nothing afterwards.
+/// exactly like the single-size probe, so a curve costs one probe per size on
+/// the first run and nothing afterwards.
 ///
 /// Native only, panics on WASM
 pub fn measure_memory_curve(client: &Client, access: MemoryAccess) -> MemoryCurve {
@@ -66,14 +44,11 @@ pub fn measure_memory_curve(client: &Client, access: MemoryAccess) -> MemoryCurv
         })
     };
 
-    cleanup_unless_pooled(client);
+    PooledProbes::cleanup_unless_held(client);
 
     MemoryCurve::new(access, points)
 }
 
-/// One measured point per size in [`working_set_sweep`], each cached exactly
-/// like the single-size probe, so a curve costs one probe per size on the
-/// first run and nothing afterwards.
 fn sweep(
     client: &Client,
     access: MemoryAccess,
@@ -121,76 +96,14 @@ pub fn measure_peak_throughput(
 
     let value = client.measure_throughput(key, || probe(client, key));
 
-    cleanup_unless_pooled(client);
+    PooledProbes::cleanup_unless_held(client);
 
     value
 }
 
-/// A device whose probes leave their pools with the allocator, for as long as
-/// one of these is alive, since faulting the next one back in costs more than
-/// the measurement. Counted per device, so a sweep that ends cannot release the
-/// pool another is still measuring against.
-struct PooledProbes {
-    service: ServiceId,
-}
-
-/// How many sweeps are holding each device's pools open.
-static POOLED_PROBES: Mutex<Option<HashMap<ServiceId, usize>>> = Mutex::new(None);
-
-impl PooledProbes {
-    fn enter(client: &Client) -> Self {
-        Self::enter_service(client.service_id())
-    }
-
-    fn enter_service(service: ServiceId) -> Self {
-        let mut pooled = POOLED_PROBES.lock();
-
-        *pooled
-            .get_or_insert_with(HashMap::new)
-            .entry(service)
-            .or_insert(0) += 1;
-
-        Self { service }
-    }
-}
-
-impl Drop for PooledProbes {
-    fn drop(&mut self) {
-        let mut pooled = POOLED_PROBES.lock();
-        let Some(sweeps) = pooled.as_mut() else {
-            return;
-        };
-
-        if let Some(holders) = sweeps.get_mut(&self.service) {
-            *holders -= 1;
-
-            if *holders == 0 {
-                sweeps.remove(&self.service);
-            }
-        }
-    }
-}
-
-/// Releases what `client` holds, unless a sweep is still measuring against it.
-/// Decided under the lock: a sweep entering between the read and the release
-/// would lose the pool it is about to measure.
-fn cleanup_unless_pooled(client: &Client) {
-    let pooled = POOLED_PROBES.lock();
-
-    if !holds(&pooled, client.service_id()) {
-        client.memory_cleanup();
-    }
-}
-
-fn holds(pooled: &Option<HashMap<ServiceId, usize>>, service: ServiceId) -> bool {
-    pooled
-        .as_ref()
-        .is_some_and(|sweeps| sweeps.contains_key(&service))
-}
-
 /// Measures `key`, in the fastest shape its probe can be launched in.
 fn probe(client: &Client, key: ThroughputKey) -> Result<ThroughputValue, ThroughputError> {
-    let launch_config = launch_config(client, key.dtype());
+    let launch_config = LaunchConfig::for_device(client, key.dtype());
 
     match key.mode {
         ThroughputMode::ComputeDirect { dtype } => {
@@ -199,60 +112,67 @@ fn probe(client: &Client, key: ThroughputKey) -> Result<ThroughputValue, Through
                 return Err(ThroughputError::Unsupported);
             }
 
-            let shapes = arithmetic_dtypes(client, dtype)
-                .into_iter()
-                .flat_map(|dtype| {
-                    arithmetic_widths(client, dtype)
-                        .into_iter()
-                        .map(move |vector_size| {
-                            (
-                                dtype,
-                                LaunchConfig {
-                                    vector_size,
-                                    ..launch_config
-                                },
-                            )
-                        })
-                })
-                .collect();
-
-            fastest_shape(shapes, |(dtype, config)| {
-                compute_direct::build_kernel(client, dtype, config)
-            })
-            .map(|(value, _)| value)
+            ShapeSweep::new(compute_direct_shapes(client, dtype, launch_config))
+                .fastest(|(dtype, config)| compute_direct::build_kernel(client, dtype, config))
+                .map(|(value, _)| value)
         }
         ThroughputMode::ComputeCmma {
             dtype,
             config: cmma_config,
         } => {
-            if !implements_cmma(client, dtype, cmma_config) {
+            if !CooperativeMatrix::implemented(client, dtype, cmma_config) {
                 return Err(ThroughputError::Unsupported);
             }
 
-            fastest_shape(alloc::vec![launch_config], |config| {
-                compute_cmma::build_kernel(client, key, cmma_config, config)
-            })
-            .map(|(value, _)| value)
+            ShapeSweep::new(alloc::vec![launch_config])
+                .fastest(|config| compute_cmma::build_kernel(client, key, cmma_config, config))
+                .map(|(value, _)| value)
         }
         ThroughputMode::Memory(spec) => {
-            let (value, fastest) = fastest_shape(
-                memory_shapes(client, launch_config, spec.access),
-                |config| match spec.access {
-                    MemoryAccess::Copy => memory_direct::build_kernel(client, key, config, spec),
-                    MemoryAccess::Read => memory_read::build_kernel(client, key, config, spec),
-                    MemoryAccess::Write => memory_write::build_kernel(client, key, config, spec),
-                },
-            )?;
+            let (value, fastest) = ShapeSweep::new(WorkerSweep::shapes(
+                client,
+                launch_config,
+                spec.access,
+            ))
+            .fastest(|config| match spec.access {
+                MemoryAccess::Copy => memory_direct::build_kernel(client, key, config, spec),
+                MemoryAccess::Read => memory_read::build_kernel(client, key, config, spec),
+                MemoryAccess::Write => memory_write::build_kernel(client, key, config, spec),
+            })?;
 
-            remember_workers(client, spec.access, fastest.cube_dim.num_elems());
+            WorkerSweep::remember(client, spec.access, fastest.cube_dim.num_elems());
 
             Ok(value)
         }
-        ThroughputMode::Launch => fastest_shape(alloc::vec![launch_config], |config| {
-            launch_overhead::build_kernel(client, key, config)
-        })
-        .map(|(value, _)| value),
+        ThroughputMode::Launch => ShapeSweep::new(alloc::vec![launch_config])
+            .fastest(|config| launch_overhead::build_kernel(client, key, config))
+            .map(|(value, _)| value),
     }
+}
+
+/// Every operand type and vector width the arithmetic ceiling for `dtype` is
+/// measured over.
+fn compute_direct_shapes(
+    client: &Client,
+    dtype: ElemType,
+    launch_config: LaunchConfig,
+) -> alloc::vec::Vec<(ElemType, LaunchConfig)> {
+    Arithmetic::dtypes(client, dtype)
+        .into_iter()
+        .flat_map(|dtype| {
+            Arithmetic::widths(client, dtype)
+                .into_iter()
+                .map(move |vector_size| {
+                    (
+                        dtype,
+                        LaunchConfig {
+                            vector_size,
+                            ..launch_config
+                        },
+                    )
+                })
+        })
+        .collect()
 }
 
 /// Calculates roofline autotune bounds for a given [`Work`] amount and compute throughput key.
@@ -289,351 +209,5 @@ pub fn roofline_bounds(
         launch_overhead: measure_peak_throughput(client, launch_key)
             .map(|value| value.duration_per_op())
             .unwrap_or_default(),
-    }
-}
-
-/// What the device answers fastest, of the shapes a probe can be launched in.
-///
-/// Shapes are built one at a time: a memory probe's pool is a large fraction of
-/// what the device will allocate. A rate that is not finite did not run.
-fn fastest_shape<S: Copy>(
-    shapes: alloc::vec::Vec<S>,
-    build: impl Fn(S) -> KernelConfig,
-) -> Result<(ThroughputValue, S), ThroughputError> {
-    let (fastest, warmed) = match shapes.len() {
-        0 => return Err(ThroughputError::NoTiming),
-        1 => (shapes[0], None),
-        _ => {
-            let (fastest, iterations) =
-                ranked_shape(&shapes, &build).ok_or(ThroughputError::NoTiming)?;
-
-            (fastest, Some(iterations))
-        }
-    };
-
-    let config = build(fastest);
-    let value = match warmed {
-        Some(iterations) => ThroughputBenchmarker::sample_at(&config, iterations),
-        None => ThroughputBenchmarker::sample(config),
-    };
-
-    value
-        .ops_per_s()
-        .is_finite()
-        .then_some((value, fastest))
-        .ok_or(ThroughputError::NoTiming)
-}
-
-/// The shape that answers fastest over a ranking pass, and the count the device
-/// was warmed at. Every shape is timed at that same count, so they are ordered
-/// on what they do rather than on which was warmed.
-fn ranked_shape<S: Copy>(shapes: &[S], build: impl Fn(S) -> KernelConfig) -> Option<(S, usize)> {
-    let mut warmed = None;
-    let mut fastest: Option<(f64, S)> = None;
-
-    for shape in shapes {
-        let config = build(*shape);
-        let start = *warmed.get_or_insert_with(|| ThroughputBenchmarker::warm(&config));
-        let ranked = ThroughputBenchmarker::rank(&config, start);
-        let rate = ranked.value.ops_per_s();
-
-        // The next shape starts from what this one settled on, so a sweep whose
-        // shapes drift in cost keeps its settling launch near the target.
-        warmed = Some(ranked.iterations);
-
-        if rate.is_finite() && fastest.is_none_or(|(best, _)| rate > best) {
-            fastest = Some((rate, *shape));
-        }
-    }
-
-    Some((fastest?.1, warmed?))
-}
-
-/// The launch shapes a memory probe is measured in. A GPU keeps the cube it was
-/// pinned to, since a wider one reports several times the bus rate. A CPU sweeps
-/// its worker count once per access and keeps the winner for every working set.
-fn memory_shapes(
-    client: &Client,
-    config: LaunchConfig,
-    access: MemoryAccess,
-) -> alloc::vec::Vec<LaunchConfig> {
-    if client.properties().hardware.num_cpu_cores.is_none() {
-        return alloc::vec![config];
-    }
-
-    if let Some(units) = remembered_workers(client, access) {
-        return alloc::vec![with_units(client, config, units)];
-    }
-
-    worker_counts(config.cube_dim.num_elems() as usize)
-        .into_iter()
-        .map(|units| with_units(client, config, units as u32))
-        .collect()
-}
-
-/// Worker counts a CPU probe is swept over, the full launch first and halvings
-/// after. Little's law would derive the count from the bandwidth, which is the
-/// thing being measured, so it is swept instead.
-fn worker_counts(units: usize) -> alloc::vec::Vec<usize> {
-    core::iter::successors(Some(units.max(1)), |units| (*units > 1).then(|| units / 2))
-        .take(MEMORY_WORKER_SHAPES)
-        .collect()
-}
-
-fn with_units(client: &Client, config: LaunchConfig, units: u32) -> LaunchConfig {
-    LaunchConfig {
-        cube_dim: CubeDim::new(client, units as usize),
-        ..config
-    }
-}
-
-/// The worker count each device streams each access fastest at, once one probe
-/// of that access has swept for it.
-static SATURATING_WORKERS: Mutex<Option<HashMap<(ServiceId, MemoryAccess), u32>>> =
-    Mutex::new(None);
-
-fn remembered_workers(client: &Client, access: MemoryAccess) -> Option<u32> {
-    let workers = SATURATING_WORKERS.lock();
-    let workers = workers.as_ref()?;
-
-    workers.get(&(client.service_id(), access)).copied()
-}
-
-fn remember_workers(client: &Client, access: MemoryAccess, units: u32) {
-    let mut workers = SATURATING_WORKERS.lock();
-
-    workers
-        .get_or_insert_with(HashMap::new)
-        .insert((client.service_id(), access), units);
-}
-
-/// The element types the arithmetic ceiling for `dtype` is measured in. Without
-/// a native fma a kernel converts and accumulates in f32 rather than pay the
-/// emulated rate, so both are measured and the faster kept.
-fn arithmetic_dtypes(client: &Client, dtype: ElemType) -> alloc::vec::Vec<ElemType> {
-    let mut dtypes = alloc::vec![dtype];
-
-    if let Some(accumulator) = promoted_dtype(dtype)
-        && client.properties().features.supports_type(accumulator)
-    {
-        dtypes.push(accumulator);
-    }
-
-    dtypes
-}
-
-/// What a kernel with `dtype` operands accumulates in when the device has no
-/// arithmetic of its own for them, or `None` where `dtype` is already the
-/// widest of the two.
-fn promoted_dtype(dtype: ElemType) -> Option<ElemType> {
-    let accumulator = ElemType::Float(FloatKind::F32);
-    let narrower_float = matches!(dtype, ElemType::Float(_)) && dtype.size() < accumulator.size();
-
-    narrower_float.then_some(accumulator)
-}
-
-/// The vector widths an arithmetic probe is measured at.
-///
-/// `io_optimized_vector_sizes` is ordered for the loads and stores this probe
-/// issues none of, and its widest is not the fastest on every device.
-fn arithmetic_widths(client: &Client, dtype: ElemType) -> alloc::vec::Vec<usize> {
-    let widths: alloc::vec::Vec<usize> = client.io_optimized_vector_sizes(dtype.size()).collect();
-
-    if widths.is_empty() {
-        alloc::vec![1]
-    } else {
-        widths
-    }
-}
-
-/// Whether the device implements the cooperative matrix the probe launches.
-///
-/// A non-empty capability list says the device has tensor hardware, not this
-/// shape of it. `mma` is not consulted: the probe issues `cmma::execute`.
-fn implements_cmma(client: &Client, dtype: ElemType, config: ComputeCmmaConfig) -> bool {
-    client.properties().features.matmul.cmma.iter().any(|it| {
-        it.a_type == dtype
-            && it.b_type == dtype
-            && it.cd_type == config.accumulator_type
-            && it.m as usize == config.cmma_dims.m
-            && it.n as usize == config.cmma_dims.n
-            && it.k as usize == config.cmma_dims.k
-    })
-}
-
-/// Hardware execution parameters for launching a compute kernel.
-#[derive(Clone, Copy)]
-pub struct LaunchConfig {
-    /// The cube the probe launches, resolved once so `ops_count` cannot
-    /// describe a launch that did not happen.
-    pub cube_dim: CubeDim,
-    /// The total number of cubes to dispatch.
-    pub cube_count: usize,
-    /// The vectorization factor (e.g., 4 for `vec4` operations).
-    pub vector_size: usize,
-    /// The number of threads in a hardware execution plane.
-    pub plane_size: usize,
-}
-
-fn launch_config(client: &Client, dtype: ElemType) -> LaunchConfig {
-    let hardware = &client.properties().hardware;
-
-    let plane_size = hardware.plane_size_max.max(1);
-    let vector_size = client
-        .io_optimized_vector_sizes(dtype.size())
-        .next()
-        .unwrap_or(1);
-
-    // A CPU has no SMs, so `sms * 32` cubes is the wrong grid to size from:
-    // a cube's units are its real dispatched workers here, while its cube
-    // count is only a loop inside each of them. `num_cpu_cores` units, one
-    // per core, is the real worker count.
-    let (units, cube_count) = match hardware.num_cpu_cores {
-        Some(cores) => (cores, CPU_CHAIN_DEPTH as u32),
-        None => {
-            let sms = hardware.num_streaming_multiprocessors.unwrap_or(64);
-            (
-                PROBE_UNITS_PER_CUBE,
-                (sms * 32).min(hardware.max_cube_count.0),
-            )
-        }
-    };
-
-    LaunchConfig {
-        cube_dim: CubeDim::new(client, units as usize),
-        cube_count: cube_count as usize,
-        vector_size,
-        plane_size: plane_size as usize,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use cubecl_common::device::DeviceId;
-    use cubecl_core::ir::{IntKind, UIntKind};
-
-    /// A service to key pooled state on, distinct per `index`.
-    fn service(index: u16) -> ServiceId {
-        ServiceId::of::<u8>(DeviceId::new(0, index))
-    }
-
-    fn pooled(service: ServiceId) -> bool {
-        holds(&POOLED_PROBES.lock(), service)
-    }
-
-    const F32: ElemType = ElemType::Float(FloatKind::F32);
-
-    /// A sweep that ends while another is running must not release the pool the
-    /// other is still measuring against, which is what a single flag would do.
-    #[test]
-    fn a_sweep_that_ends_leaves_an_overlapping_one_pooled() {
-        let device = service(1);
-        let held = PooledProbes::enter_service(device);
-
-        {
-            let _ends_first = PooledProbes::enter_service(device);
-            assert!(pooled(device));
-        }
-
-        assert!(pooled(device), "one sweep is still running");
-        drop(held);
-        assert!(!pooled(device));
-    }
-
-    /// The same, across threads, where the two sweeps genuinely overlap rather
-    /// than nest.
-    #[test]
-    fn a_sweep_on_another_thread_keeps_its_own_device_pooled() {
-        use std::sync::{Arc, Barrier};
-
-        let device = service(2);
-        let (entered, ended) = (Arc::new(Barrier::new(2)), Arc::new(Barrier::new(2)));
-
-        let holder = std::thread::spawn({
-            let (entered, ended) = (entered.clone(), ended.clone());
-
-            move || {
-                let _pooled = PooledProbes::enter_service(device);
-                entered.wait();
-                ended.wait();
-            }
-        });
-
-        {
-            let _pooled = PooledProbes::enter_service(device);
-            entered.wait();
-        }
-
-        assert!(pooled(device), "the other thread's sweep is still running");
-        ended.wait();
-        holder.join().expect("the holding thread finishes");
-        assert!(!pooled(device));
-    }
-
-    /// Pools are a device's own, and two cards of the same model are two
-    /// devices: a sweep on one says nothing about the other.
-    #[test]
-    fn a_sweep_pools_only_the_device_it_measures() {
-        let _pooled = PooledProbes::enter_service(service(3));
-
-        assert!(pooled(service(3)));
-        assert!(!pooled(service(4)));
-    }
-
-    /// The reported value is a full measurement of the winner, not the
-    /// ranking pass that found it.
-    #[test]
-    fn the_shape_that_ranks_fastest_is_the_one_measured() {
-        // A shape here is its rate: the faster it is, the less time a pass takes.
-        let build = |rate: u64| KernelConfig {
-            sample: alloc::boxed::Box::new(move |iterations| {
-                core::time::Duration::from_nanos(iterations as u64 * 1000 / rate)
-            }),
-            ops_count: 1,
-            min_iterations: 1,
-        };
-
-        let (value, fastest) = fastest_shape(alloc::vec![1, 4, 2], build).expect("a shape ran");
-
-        assert_eq!(fastest, 4);
-        assert!(value.ops_per_s().is_finite());
-    }
-
-    /// The full launch is measured first, so a device whose peak is there
-    /// keeps the shape it had before the sweep existed.
-    #[test]
-    fn the_sweep_starts_at_the_full_launch_and_halves_to_the_budget() {
-        assert_eq!(worker_counts(16), alloc::vec![16, 8, 4, 2, 1]);
-        assert_eq!(worker_counts(128), alloc::vec![128, 64, 32, 16, 8]);
-    }
-
-    /// Every count is a launch, so one core must not be handed an empty sweep
-    /// and reported as untimeable.
-    #[test]
-    fn one_core_is_still_a_shape() {
-        assert_eq!(worker_counts(1), alloc::vec![1]);
-    }
-
-    /// Where the device does the arithmetic itself, its rate is the ceiling and
-    /// a second measurement of the same thing costs a probe for nothing.
-    #[test]
-    fn nothing_as_wide_as_the_accumulator_is_promoted() {
-        assert_eq!(promoted_dtype(F32), None);
-        assert_eq!(promoted_dtype(ElemType::Float(FloatKind::F64)), None);
-    }
-
-    /// The probe retires a multiply for an integer, which no float rate bounds.
-    #[test]
-    fn an_integer_is_not_promoted() {
-        assert_eq!(promoted_dtype(ElemType::Int(IntKind::I8)), None);
-        assert_eq!(promoted_dtype(ElemType::UInt(UIntKind::U16)), None);
-    }
-
-    #[test]
-    fn every_float_narrower_than_the_accumulator_is_promoted() {
-        for dtype in [FloatKind::F16, FloatKind::BF16, FloatKind::E4M3] {
-            assert_eq!(promoted_dtype(ElemType::Float(dtype)), Some(F32));
-        }
     }
 }
