+++ crates/cubecl-std/src/throughput/mod.rs
@@ -1,5 +1,16 @@
 mod base;
+mod launch;
+mod operands;
+mod pooling;
 mod runners;
+mod shape;
+mod workers;
 
 pub use base::*;
+pub use launch::*;
 pub use runners::*;
+
+use operands::{Arithmetic, CooperativeMatrix};
+use pooling::PooledProbes;
+use shape::ShapeSweep;
+use workers::WorkerSweep;
